Envip da ordem para producao e recebimento de StatusPedido

+++ src/main/java/br/com/postech/senderorder/sevenfoodorderapi/core/domain/OrderStatusDTO.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class OrderStatusDTO {
     private String orderId;
-    private Long statusPedido;
+    private Integer statusPedido;
 }
